server/pyproject.toml
[project]
name = "seeme-server"
version = "0.1.0"
description = "seeme-server"
authors = [
    {name = "lolik",email = "tuwork@foxmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10,<4"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "mcstatus (>=12.00,<=12.0.6)",
    "redis (>=5.2.1,<6.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "requests (>=2.32.3,<3.0.0)",
    "openai (>=1.65.2,<2.0.0)"
]

[dependency-groups]
dev = [
    "fabric>=3.2.2",
    "pre-commit>=4.4.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.5",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 200
exclude = [".venv"]
